Added Credits to MainMenu

+++ TempleOfDoom/TempleOfDoom/Character.cpp
@@ -24,7 +24,7 @@ namespace TOD {
 		//// Calculate damage
 		//int damage = totaldefense - attack < 0 ? (totaldefense - attack) * -1 : 0;
 
-		int multiplier = (level - c->getLevel()) * 0.4; // Damage multiplier
+		int multiplier = (int)(level - c->getLevel()) * 0.4; // Damage multiplier
 		int damage = 1;
 		damage += (c->getDefense() - attack) * -1;
